+++ src/device/include/TaskHC_SR04.h
@@ -5,6 +5,6 @@
 
 extern void TaskHC_SR04(void *pvParameters);
 extern void initHC_SR04();
-extern int getDistanceHC_SR04();
+extern float getDistanceHC_SR04();
 
 #endif /* INC_TASKHC_SR04_H_ */
